Migrating more classes to kotlin

+++ mbrc-lib/src/main/kotlin/com/kelsos/mbrc/dto/library/AlbumDto.kt
@@ -10,60 +10,19 @@ import com.kelsos.mbrc.empty
 @JsonPropertyOrder("name", "artist_id", "cover_id", "album_id", "id", "date_added", "date_updated", "date_deleted")
 class AlbumDto {
 
-    /**
-     * @return The name
-     */
-    /**
-     * @param name The name
-     */
     @JsonProperty("name") var name: String = String.empty
-    /**
-     * @return The artistId
-     */
-    /**
-     * @param artistId The artist_id
-     */
-    @JsonProperty("artist_id") var artistId: Int = 0
-    /**
-     * @return The coverId
-     */
-    /**
-     * @param coverId The cover_id
-     */
-    @JsonProperty("cover_id") var coverId: Int = 0
-    /**
-     * @return The albumId
-     */
-    /**
-     * @param albumId The album_id
-     */
+
+    @JsonProperty("artist_id") var artistId: Long = 0
+
+    @JsonProperty("cover_id") var coverId: Long = 0
+
     @JsonProperty("album_id") var albumId: String = String.empty
-    /**
-     * @return The id
-     */
-    /**
-     * @param id The id
-     */
-    @JsonProperty("id") var id: Int = 0
-    /**
-     * @return The dateAdded
-     */
-    /**
-     * @param dateAdded The date_added
-     */
+
+    @JsonProperty("id") var id: Long = 0
+
     @JsonProperty("date_added") var dateAdded: Long = 0
-    /**
-     * @return The dateUpdated
-     */
-    /**
-     * @param dateUpdated The date_updated
-     */
+
     @JsonProperty("date_updated") var dateUpdated: Long = 0
-    /**
-     * @return The dateDeleted
-     */
-    /**
-     * @param dateDeleted The date_deleted
-     */
+
     @JsonProperty("date_deleted") var dateDeleted: Long = 0
 }
